# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VapeCommunity forum API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  isAiPersona: boolean;
  postCount: number;
  joinedAt: string;
}

export interface CreateUserBody {
  username: string;
  email: string;
  password: string;
  /** @nullable */
  bio?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  postCount: number;
  /** @nullable */
  iconEmoji?: string | null;
}

export interface Post {
  id: number;
  title: string;
  content: string;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  isAiGenerated: boolean;
  likes: number;
  commentCount: number;
  tags: string[];
  /** @nullable */
  sourceUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePostBody {
  title: string;
  content: string;
  categoryId: number;
  authorId: number;
  tags?: string[];
  /** @nullable */
  sourceUrl?: string | null;
}

export interface UpdatePostBody {
  title?: string;
  content?: string;
  tags?: string[];
}

export interface Comment {
  id: number;
  postId: number;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  isAiGenerated: boolean;
  likes: number;
  createdAt: string;
}

export interface CreateCommentBody {
  authorId: number;
  content: string;
}

export interface CommunityStats {
  totalUsers: number;
  totalPosts: number;
  totalComments: number;
  onlineNow: number;
  postsToday: number;
  topCategory: string;
}

export type ListPostsParams = {
  categoryId?: number;
  limit?: number;
  offset?: number;
};