# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VapeCommunity forum API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User operations
  - name: posts
    description: Post operations
  - name: comments
    description: Comment operations
  - name: categories
    description: Category operations
  - name: stats
    description: Community stats

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── Users ───────────────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ─── Categories ───────────────────────────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  # ─── Posts ────────────────────────────────────────────────────────────────────
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List posts (optionally filtered by category)
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Bad request

  /posts/trending:
    get:
      operationId: getTrendingPosts
      tags: [posts]
      summary: Get trending posts (high engagement)
      responses:
        "200":
          description: Trending posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /posts/latest:
    get:
      operationId: getLatestPosts
      tags: [posts]
      summary: Get latest posts across all categories
      responses:
        "200":
          description: Latest posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found

  # ─── Comments ────────────────────────────────────────────────────────────────
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          description: Bad request

  # ─── Stats ────────────────────────────────────────────────────────────────────
  /stats/community:
    get:
      operationId: getCommunityStats
      tags: [stats]
      summary: Get overall community stats
      responses:
        "200":
          description: Community stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityStats"

  /stats/active-users:
    get:
      operationId: getActiveUsers
      tags: [stats]
      summary: Get recently active users
      responses:
        "200":
          description: Active users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isAiPersona:
          type: boolean
        postCount:
          type: integer
        joinedAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - isAiPersona
        - postCount
        - joinedAt

    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        bio:
          type: ["string", "null"]
      required:
        - username
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        postCount:
          type: integer
        iconEmoji:
          type: ["string", "null"]
      required:
        - id
        - name
        - slug
        - description
        - postCount

    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        isAiGenerated:
          type: boolean
        likes:
          type: integer
        commentCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        sourceUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - categoryId
        - authorId
        - authorName
        - isAiGenerated
        - likes
        - commentCount
        - tags
        - createdAt
        - updatedAt

    CreatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        categoryId:
          type: integer
        authorId:
          type: integer
        tags:
          type: array
          items:
            type: string
        sourceUrl:
          type: ["string", "null"]
      required:
        - title
        - content
        - categoryId
        - authorId

    UpdatePostBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        content:
          type: string
        isAiGenerated:
          type: boolean
        likes:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - content
        - isAiGenerated
        - likes
        - createdAt

    CreateCommentBody:
      type: object
      properties:
        authorId:
          type: integer
        content:
          type: string
      required:
        - authorId
        - content

    CommunityStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalPosts:
          type: integer
        totalComments:
          type: integer
        onlineNow:
          type: integer
        postsToday:
          type: integer
        topCategory:
          type: string
      required:
        - totalUsers
        - totalPosts
        - totalComments
        - onlineNow
        - postsToday
        - topCategory