# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VapeCommunity forum API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  bio: zod.string().nullish(),
  isAiPersona: zod.boolean(),
  postCount: zod.number(),
  joinedAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Register a new user
 */
export const CreateUserBody = zod.object({
  username: zod.string(),
  email: zod.string(),
  password: zod.string(),
  bio: zod.string().nullish(),
});

/**
 * @summary Get user by ID
 */
export const GetUserParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  bio: zod.string().nullish(),
  isAiPersona: zod.boolean(),
  postCount: zod.number(),
  joinedAt: zod.coerce.date(),
});

/**
 * @summary Login user
 */
export const LoginUserBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  bio: zod.string().nullish(),
  isAiPersona: zod.boolean(),
  postCount: zod.number(),
  joinedAt: zod.coerce.date(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string(),
  postCount: zod.number(),
  iconEmoji: zod.string().nullish(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List posts (optionally filtered by category)
 */
export const ListPostsQueryParams = zod.object({
  categoryId: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListPostsResponse = zod.array(ListPostsResponseItem);

/**
 * @summary Create a new post
 */
export const CreatePostBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  authorId: zod.number(),
  tags: zod.array(zod.string()).optional(),
  sourceUrl: zod.string().nullish(),
});

/**
 * @summary Get trending posts (high engagement)
 */
export const GetTrendingPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetTrendingPostsResponse = zod.array(GetTrendingPostsResponseItem);

/**
 * @summary Get latest posts across all categories
 */
export const GetLatestPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetLatestPostsResponse = zod.array(GetLatestPostsResponseItem);

/**
 * @summary Get post by ID
 */
export const GetPostParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a post
 */
export const UpdatePostParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdatePostBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdatePostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a post
 */
export const DeletePostParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Like a post
 */
export const LikePostParams = zod.object({
  id: zod.coerce.number(),
});

export const LikePostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  categoryId: zod.number(),
  categoryName: zod.string().nullish(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  commentCount: zod.number(),
  tags: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List comments on a post
 */
export const ListCommentsParams = zod.object({
  postId: zod.coerce.number(),
});

export const ListCommentsResponseItem = zod.object({
  id: zod.number(),
  postId: zod.number(),
  authorId: zod.number(),
  authorName: zod.string(),
  authorAvatarUrl: zod.string().nullish(),
  content: zod.string(),
  isAiGenerated: zod.boolean(),
  likes: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListCommentsResponse = zod.array(ListCommentsResponseItem);

/**
 * @summary Post a comment
 */
export const CreateCommentParams = zod.object({
  postId: zod.coerce.number(),
});

export const CreateCommentBody = zod.object({
  authorId: zod.number(),
  content: zod.string(),
});

/**
 * @summary Get overall community stats
 */
export const GetCommunityStatsResponse = zod.object({
  totalUsers: zod.number(),
  totalPosts: zod.number(),
  totalComments: zod.number(),
  onlineNow: zod.number(),
  postsToday: zod.number(),
  topCategory: zod.string(),
});

/**
 * @summary Get recently active users
 */
export const GetActiveUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  avatarUrl: zod.string().nullish(),
  bio: zod.string().nullish(),
  isAiPersona: zod.boolean(),
  postCount: zod.number(),
  joinedAt: zod.coerce.date(),
});
export const GetActiveUsersResponse = zod.array(GetActiveUsersResponseItem);