# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VapeCommunity forum API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  postId: number;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatarUrl?: string | null;
  content: string;
  isAiGenerated: boolean;
  likes: number;
  createdAt: Date;
}