# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VapeCommunity forum API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentBody {
  authorId: number;
  content: string;
}