# lib/api-zod/src/generated/types/createPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VapeCommunity forum API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePostBody {
  title: string;
  content: string;
  categoryId: number;
  authorId: number;
  tags?: string[];
  /** @nullable */
  sourceUrl?: string | null;
}